src/main/java/com/estudo/collection/set/ordenacao/exercicio01/CadastroProdutos.java
package com.estudo.collection.set.ordenacao.exercicio01;

import java.util.HashSet;
import java.util.Set;
import java.util.TreeSet;

public class CadastroProdutos {
    private Set<Produto> conjuntoProduto;

    public CadastroProdutos() {
        this.conjuntoProduto = new HashSet<>();
    }

    public void adicionarProduto(long codigo, String nome, double preco, int quantidade) {
        this.conjuntoProduto.add(new Produto(nome, codigo, preco, quantidade));
    }

    public Set<Produto> exibirProdutosPorNome() {
        Set<Produto> produtosPorNome = new TreeSet<>(conjuntoProduto);
        return produtosPorNome;
    }

    public Set<Produto> exibirProdutosPorPreco() {
        Set<Produto> produtosPorPreco = new TreeSet<>(new ComparatorPorPreco());
        produtosPorPreco.addAll(conjuntoProduto);
        return produtosPorPreco;
    }
}
